requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
black==24.4.2
certifi==2025.6.15
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
dnspython==2.6.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.110.1
flake8==7.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
motor==3.3.2
mypy-extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pyflakes==3.2.0
pymongo==4.10.1
pytest==8.2.2
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tomli==2.2.1
typing-extensions==4.13.2
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
